+++ ProjectAI/src/Action/ICanvasAction.java
@@ -33,14 +33,14 @@ public abstract class ICanvasAction extends Canvas {
 
 	public void resizeParent() {
 		if (action.isShownCond()) {
-			int x1 = action.getLengthPrec() + action.getLengthEff() + action.getWidthRect();
-			int y1 = action.getHeightRect() + 40;
-			parent.setSize(x1, y1);
+			double x1 = action.getLengthPrec() + action.getLengthEff() + action.getWidthRect();
+			double y1 = action.getHeightRect() + 40;
+			parent.setSize((int)x1,(int) y1);
 
 		} else {
-			int x1 = action.getStandardLengthPrec() + action.getStandardLengthEff() + action.getWidthRect();
-			int y1 = action.getHeightRect() + 40;
-			parent.setSize(x1, y1);
+			double x1 = action.getStandardLengthPrec() + action.getStandardLengthEff() + action.getWidthRect();
+			double y1 = action.getHeightRect() + 40;
+			parent.setSize((int)x1, (int)y1);
 		}
 	}
